Use WeakMap instead of a plain object for storage
can-dom-data now uses a WeakMap with elements as keys and plain objects as values to store the key/value pairs that are set.

WeakMap guarantees that when there are no other references to its keys (elements in this case), then the values will be garbage collected, which means can-dom-data does not need to track when the DOM is mutated and remove the data it stores about elements.

Part of https://github.com/canjs/canjs/issues/4044